separate hello test
